backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
telethon
python-dotenv
requests
passlib[bcrypt]
python-multipart
python-jose[cryptography]
httpx
